Use extensions to start/stop vpn
