+++ app/src/main/java/com/betterfilter/AutoRestartActivity.kt
@@ -4,7 +4,10 @@ import android.content.Intent
 import android.net.VpnService
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.betterfilter.Extensions.startVpn
 import com.betterfilter.vpn.VpnHostsService
+import com.betterfilter.vpn.vpn.AdVpnService
+import com.betterfilter.vpn.vpn.Command
 
 class AutoRestartActivity : AppCompatActivity() {
 
@@ -13,9 +16,7 @@ class AutoRestartActivity : AppCompatActivity() {
         setContentView(R.layout.activity_auto_restart)
 
         if (!intent.getBooleanExtra("isFromOurButton", false)) {
-            VpnService.prepare(this)
-            val intent = Intent(this, VpnHostsService::class.java)
-            startService(intent)
+            this.startVpn()
         }
         finish()
